crocodile: support booting from FlexSPI NOR flash
Extend crocodile_fspi_defconfig with configurations for booting from
FlexSPI NOR flash and provide node for binman in order to add the
default FlexSPI configuration block to the image.

The configuration block will be replaced with a proper method to
generate it, but as discussions are ongoing we can use the blob for now.
The blob is taken from imx-mkimage, where it is generated with `xxd`
from a text file. Detailed information about FlexSPI configuration block
can be found in i.MX 8M Mini Applications Processor Reference Manual.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>